requirements.txt
funcy
mock
ruamel.yaml
pyyaml